Fix search bar hiding result sources (#828)
* Remove css hiding search sources; convert TypeaheadMenuItem to functional component in TSX

* Remove unused CSS

* switch test file to tsx

* Adjust tests for typeahead; remove function test, leave side effect test - createDescription
