+++ eventkit_cloud/ui/static/ui/app/styles/typeaheadStyles.css
@@ -44,10 +44,3 @@ has no support for inline styles :(
 .menuItem > a {
     padding: 7px 10px !important;
 }
-
-/*
-This hides an unused label on the right side of result items.
- */
-.row > div > strong {
-    display: none !important;
-}
